Left bug in debug statement. Oops.

+++ models/siamese_model.py
@@ -49,8 +49,8 @@ class Siamese_Model(Model):
 
             print("Expected:",expected)
             print("preds:", preds)
-            print("equivalency of exp and preds:", preds.float() == expected)
-            print("Sum of above:", torch.sum(preds.float() == expected))
+            print("equivalency of exp and preds:", preds == expected)
+            print("Sum of above:", torch.sum(preds == expected))
             #print("Shape of preds:",preds.shape)
             #print("Shape of labels:", expected.shape)
             #print("Number of labels:", len(labels))
